Move jit dev compute calculation to startup
